separated server from client
